Fix schema check for predefined, acceptable values
Fixes the check in `illFormedOpts` for predefined, acceptable values in
the provided `options` schema. Node (v8) recently added support for the
`Object.values()` method, which `illFormedOpts ` was incorrectly
treating as an `Object` literal with the property `values` (mapping to
an array of predefined acceptable values). This fix amends the check to
ensure the schema property is an `Object` literal.